edit

+++ murmurhash_util.cpp
@@ -20,12 +20,12 @@
 
 // 64-bit hash for 64-bit platforms
 
-uint32_t MurmurHash64A ( const void * key, int len, unsigned int seed )
+uint64_t MurmurHash64A ( const void * key, int len, uint64_t seed )
 {
   const uint64_t m = 0xc6a4a7935bd1e995;
   const int r = 47;
 
-  uint32_t h = seed ^ (len * m);
+  uint64_t h = seed ^ (len * m);
 
   const uint64_t * data = (const uint64_t *)key;
   const uint64_t * end = data + (len/8);
@@ -66,7 +66,7 @@ uint32_t MurmurHash64A ( const void * key, int len, unsigned int seed )
 
 // 64-bit hash for 32-bit platforms
 
-uint32_t MurmurHash64B ( const void * key, int len, unsigned int seed )
+uint64_t MurmurHash64B ( const void * key, int len, unsigned int seed )
 {
   const unsigned int m = 0x5bd1e995;
   const int r = 24;
@@ -110,7 +110,7 @@ uint32_t MurmurHash64B ( const void * key, int len, unsigned int seed )
   h1 ^= h2 >> 17; h1 *= m;
   h2 ^= h1 >> 19; h2 *= m;
 
-  uint32_t h = h1;
+  uint64_t h = h1;
 
   h = (h << 32) | h2;
 
@@ -129,7 +129,7 @@ uint32_t MurmurHash64B ( const void * key, int len, unsigned int seed )
 // Thomas Wang's integer hash functions. See
 // <https://gist.github.com/lh3/59882d6b96166dfc3d8d> for a snapshot.
 
-uint32_t hash_64(uint64_t key, uint64_t mask)
+uint64_t hash_64(uint64_t key, uint64_t mask)
 {
   key = (~key + (key << 21)) & mask; // key = (key << 21) - key - 1;
   key = key ^ key >> 24;
@@ -143,9 +143,9 @@ uint32_t hash_64(uint64_t key, uint64_t mask)
 
 // The inversion of hash_64(). Modified from
 // <https://naml.us/blog/tag/invertible>
-uint32_t hash_64i(uint64_t key, uint64_t mask)
+uint64_t hash_64i(uint64_t key, uint64_t mask)
 {
-  uint32_t tmp;
+  uint64_t tmp;
 
   // Invert key = key + (key << 31)
   tmp = (key - (key << 31));
